add wifi manager

+++ src/src/tools.cpp
@@ -1,10 +1,10 @@
 #include <tools.h>
 #include "constants_defines.h"
 
-int tools::ctoi(char x){
+ short int tools::ctoi(char x){
 	return x - '0';
 }
-int tools::timetototalSeconds(String time){
+ long tools::timetototalSeconds(String time){
   int hours = ctoi(time[0])*10+ctoi(time[1]);
   int minutes = ctoi(time[3])*10+ctoi(time[4]);
   return hours*60*60+minutes*60;
